Move custom serialization into different file
