+++ event-models/rust/src/main.rs
@@ -1,7 +1,8 @@
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let event = fnhs_event_models::parse()?;
     println!("Event: {:?}", event);
-    let event = fnhs_event_models::custom_parse()?;
+    fnhs_event_models::custom::test_stringify()?;
+    let event = fnhs_event_models::custom::test_parse()?;
     println!("Custom Event: {:?}", event);
     Ok(())
 }
